+++ lib/seo_widget/seo_bottomnavigationbar.dart
@@ -1,0 +1,112 @@
+import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_dom_ui/helper/seo_injectable.dart';
+import 'package:flutter_dom_ui/dom_injector.dart';
+
+class SeoBottomNavigationBar extends StatelessWidget
+    implements SeoInjectableLayout {
+  final List<BottomNavigationBarItem> items;
+  final void Function(int)? onTap;
+  final int currentIndex;
+  final double? elevation;
+  final BottomNavigationBarType? type;
+  final Color? fixedColor;
+  final Color? backgroundColor;
+  final double iconSize;
+  final Color? selectedItemColor;
+  final Color? unselectedItemColor;
+  final IconThemeData? selectedIconTheme;
+  final IconThemeData? unselectedIconTheme;
+  final double selectedFontSize;
+  final double unselectedFontSize;
+  final TextStyle? selectedLabelStyle;
+  final TextStyle? unselectedLabelStyle;
+  final bool? showSelectedLabels;
+  final bool? showUnselectedLabels;
+  final MouseCursor? mouseCursor;
+  final bool? enableFeedback;
+  final BottomNavigationBarLandscapeLayout? landscapeLayout;
+  final bool useLegacyColorScheme;
+
+  const SeoBottomNavigationBar({
+    super.key,
+    required this.items,
+    this.onTap,
+    this.currentIndex = 0,
+    this.elevation,
+    this.type,
+    this.fixedColor,
+    this.backgroundColor,
+    this.iconSize = 24.0,
+    this.selectedItemColor,
+    this.unselectedItemColor,
+    this.selectedIconTheme,
+    this.unselectedIconTheme,
+    this.selectedFontSize = 14.0,
+    this.unselectedFontSize = 12.0,
+    this.selectedLabelStyle,
+    this.unselectedLabelStyle,
+    this.showSelectedLabels,
+    this.showUnselectedLabels,
+    this.mouseCursor,
+    this.enableFeedback,
+    this.landscapeLayout,
+    this.useLegacyColorScheme = true,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return BottomNavigationBar(
+      items: items,
+      onTap: onTap,
+      currentIndex: currentIndex,
+      elevation: elevation,
+      backgroundColor: backgroundColor,
+      iconSize: iconSize,
+      selectedItemColor: selectedItemColor,
+      unselectedItemColor: unselectedItemColor,
+      selectedIconTheme: selectedIconTheme,
+      unselectedIconTheme: unselectedIconTheme,
+      selectedFontSize: selectedFontSize,
+      unselectedFontSize: unselectedFontSize,
+      selectedLabelStyle: selectedLabelStyle,
+      unselectedLabelStyle: unselectedLabelStyle,
+      showSelectedLabels: showSelectedLabels,
+      showUnselectedLabels: showUnselectedLabels,
+      mouseCursor: mouseCursor,
+      enableFeedback: enableFeedback,
+      landscapeLayout: landscapeLayout,
+      useLegacyColorScheme: useLegacyColorScheme,
+    );
+  }
+
+  @override
+  void injectHtmlTo(WebHTMLElement parent) {
+    if (!kIsWeb) return;
+    final document = webWindow.document;
+    final footerContainer = document.createElement('footer') as WebHTMLElement;
+
+    // Actions
+    if (items.isNotEmpty) {
+      final actionsContainer = document.createElement('div') as WebHTMLElement;
+      for (final item in items) {
+        _appendWidgetToContainer(item, actionsContainer);
+      }
+      footerContainer.appendChild(actionsContainer);
+    }
+
+    parent.appendChild(footerContainer);
+  }
+
+  void _appendWidgetToContainer(
+    BottomNavigationBarItem widget,
+    WebHTMLElement container,
+  ) {
+    if (widget is SeoInjectable) {
+      final el = (widget as SeoInjectable).createHtmlElement();
+      if (el != null) container.appendChild(el);
+    } else if (widget is SeoInjectableLayout) {
+      (widget as SeoInjectableLayout).injectHtmlTo(container);
+    }
+  }
+}